A commit for review, there's some unclear counters code. Also added a more precise NTSC vsync rate.
git-svn-id: http://pcsx2-playground.googlecode.com/svn/trunk@326 a6443dda-0b58-4228-96e9-037be469359c

+++ pcsx2/Misc.h
@@ -123,8 +123,8 @@ typedef struct {
 	int Hacks;
 	int GameFixes;
 	int CustomFrameSkip;
-	int CustomConsecutiveFrames;
-	int CustomConsecutiveSkip;
+	u8 CustomConsecutiveFrames;
+	u8 CustomConsecutiveSkip;
 	u32 sseMXCSR;
 	u32 sseVUMXCSR;
 } PcsxConfig;
